Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@12053 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/vfscache.cpp
@@ -61,7 +61,7 @@ iVFS* csVfsCacheManager::GetVFS ()
 void csVfsCacheManager::CacheName (char* buf, const char* type,
 	const char* scope, uint32 id)
 {
-  if (id == ~0)
+  if (id == (uint32)~0)
   {
     if (scope == NULL)
       sprintf (buf, "%s", type);
